Fixes fixedColor/selectedItemColor bug
If `selectedItemColor` was set, then `BottomNavigationBar` would set both `selectedItemColor` and `fixedColor` to `selectedItemColor`.

When it came time to create the copied `BottomNavigationBar`, this code would set both `selectedItemColor` and `fixedColor`, which would guaranteed to both be non-null, and thus would trigger an assertion error inside of `BottomNavigationBar`.

The fix is to not set `fixedColor` when `selectedItemColor` is set.